Introduce light node & refactor RPC client (#681)
* introduce light node
* add config option to start light node
* refactor `rpc/server.go/Server` to accept the `tendermint/rpc/client` interface

Closes #662; Closes #663; Closes #664; Closes #682.
