+++ config/defaults.go
@@ -26,4 +26,5 @@ var DefaultNodeConfig = NodeConfig{
 	},
 	DALayer:  "mock",
 	DAConfig: "",
+	Light:    false,
 }
