+++ rpc/json/service.go
@@ -9,17 +9,14 @@ import (
 	"time"
 
 	"github.com/gorilla/rpc/v2/json2"
-	"github.com/tendermint/tendermint/libs/pubsub"
-	tmquery "github.com/tendermint/tendermint/libs/pubsub/query"
 	rpcclient "github.com/tendermint/tendermint/rpc/client"
 	ctypes "github.com/tendermint/tendermint/rpc/core/types"
 
 	"github.com/celestiaorg/rollmint/log"
-	"github.com/celestiaorg/rollmint/rpc/client"
 )
 
 // GetHTTPHandler returns handler configured to serve Tendermint-compatible RPC.
-func GetHTTPHandler(l *client.Client, logger log.Logger) (http.Handler, error) {
+func GetHTTPHandler(l rpcclient.Client, logger log.Logger) (http.Handler, error) {
 	return newHandler(newService(l, logger), json2.NewCodec(), logger), nil
 }
 
@@ -42,12 +39,12 @@ func newMethod(m interface{}) *method {
 }
 
 type service struct {
-	client  *client.Client
+	client  rpcclient.Client
 	methods map[string]*method
 	logger  log.Logger
 }
 
-func newService(c *client.Client, l log.Logger) *service {
+func newService(c rpcclient.Client, l log.Logger) *service {
 	s := service{
 		client: c,
 		logger: l,
@@ -87,51 +84,31 @@ func newService(c *client.Client, l log.Logger) *service {
 }
 
 func (s *service) Subscribe(req *http.Request, args *subscribeArgs, wsConn *wsConn) (*ctypes.ResultSubscribe, error) {
-	addr := req.RemoteAddr
-
 	// TODO(tzdybal): pass config and check subscriptions limits
-
-	q, err := tmquery.New(args.Query)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse query: %w", err)
-	}
-
-	s.logger.Debug("subscribe to query", "remote", addr, "query", args.Query)
-
 	// TODO(tzdybal): extract consts or configs
 	const SubscribeTimeout = 5 * time.Second
 	const subBufferSize = 100
+
+	addr := req.RemoteAddr
+	query := args.Query
+
 	ctx, cancel := context.WithTimeout(req.Context(), SubscribeTimeout)
 	defer cancel()
 
-	sub, err := s.client.EventBus.Subscribe(ctx, addr, q, subBufferSize)
+	sub, err := s.client.Subscribe(ctx, addr, query, subBufferSize)
 	if err != nil {
 		return nil, fmt.Errorf("failed to subscribe: %w", err)
 	}
 
 	go func() {
-		for {
-			select {
-			case msg := <-sub.Out():
-				data, err := json.Marshal(msg.Data())
-				if err != nil {
-					s.logger.Error("failed to marshal response data", "error", err)
-					continue
-				}
-				if wsConn != nil {
-					wsConn.queue <- data
-				}
-			case <-sub.Cancelled():
-				if sub.Err() != pubsub.ErrUnsubscribed {
-					var reason string
-					if sub.Err() == nil {
-						reason = "unknown failure"
-					} else {
-						reason = sub.Err().Error()
-					}
-					s.logger.Error("subscription was cancelled", "reason", reason)
-				}
-				return
+		for msg := range sub {
+			data, err := json.Marshal(msg.Data)
+			if err != nil {
+				s.logger.Error("failed to marshal response data", "error", err)
+				continue
+			}
+			if wsConn != nil {
+				wsConn.queue <- data
 			}
 		}
 	}()
