+++ rpc/json/service_test.go
@@ -22,12 +22,12 @@ import (
 	abciclient "github.com/tendermint/tendermint/abci/client"
 	abci "github.com/tendermint/tendermint/abci/types"
 	"github.com/tendermint/tendermint/libs/log"
+	rpcclient "github.com/tendermint/tendermint/rpc/client"
 	"github.com/tendermint/tendermint/types"
 
 	"github.com/celestiaorg/rollmint/config"
 	"github.com/celestiaorg/rollmint/mocks"
 	"github.com/celestiaorg/rollmint/node"
-	"github.com/celestiaorg/rollmint/rpc/client"
 )
 
 func TestHandlerMapping(t *testing.T) {
@@ -269,7 +269,7 @@ func TestSubscription(t *testing.T) {
 }
 
 // copied from rpc
-func getRPC(t *testing.T) (*mocks.Application, *client.Client) {
+func getRPC(t *testing.T) (*mocks.Application, rpcclient.Client) {
 	t.Helper()
 	require := require.New(t)
 	app := &mocks.Application{}
@@ -292,14 +292,14 @@ func getRPC(t *testing.T) (*mocks.Application, *client.Client) {
 	})
 	key, _, _ := crypto.GenerateEd25519Key(rand.Reader)
 	signingKey, _, _ := crypto.GenerateEd25519Key(rand.Reader)
-	node, err := node.NewNode(context.Background(), config.NodeConfig{Aggregator: true, DALayer: "mock", BlockManagerConfig: config.BlockManagerConfig{BlockTime: 1 * time.Second}}, key, signingKey, abciclient.NewLocalClient(nil, app), &types.GenesisDoc{ChainID: "test"}, log.TestingLogger())
+	n, err := node.NewNode(context.Background(), config.NodeConfig{Aggregator: true, DALayer: "mock", BlockManagerConfig: config.BlockManagerConfig{BlockTime: 1 * time.Second}, Light: false}, key, signingKey, abciclient.NewLocalClient(nil, app), &types.GenesisDoc{ChainID: "test"}, log.TestingLogger())
 	require.NoError(err)
-	require.NotNil(node)
+	require.NotNil(n)
 
-	err = node.Start()
+	err = n.Start()
 	require.NoError(err)
 
-	local := client.NewClient(node)
+	local := n.GetClient()
 	require.NotNil(local)
 
 	return app, local
